Prepare PR #80 merge
Thanks @johnmikep!

New configuration dict:

"GoogleChatNotification": {
  "backend": "zentral.core.actions.backends.google_chat",
  "webhook": "https://chat.googleapis.com/v1/spaces/…"
}

See https://developers.google.com/hangouts/chat/how-tos/webhooks
